+++ src/rendering/Camera.cpp
@@ -5,7 +5,7 @@
 
 Camera::Camera(float aspectRatio)
 {
-    projection = Matrix4f::fov_perspective(100.0f, 10000.f, 45.f * M_PI / 180.f, aspectRatio);
+    projection = Matrix4f::fov_perspective(100.0f, 10000.f, 45.f * (float)M_PI / 180.f, aspectRatio);
 }
 
 Matrix4f Camera::getProjection() const
